Users can only view their own objects

+++ server/controllers/contact.js
@@ -4,7 +4,7 @@ import ContactModel from '../models/contacts.js';
 // Get All Contacts = Same as db.contacts.find()
 export const getAllContacts = async (req, res) => {
     try {
-        const contacts = await ContactModel.find();
+        const contacts = await ContactModel.find({ user: req.user.id });
         res.status(200).json(contacts);
     } catch (error) {
         // 500 HTTP status code for server error
@@ -15,10 +15,13 @@ export const getAllContacts = async (req, res) => {
 // Read a contact by ID = Same as db.contacts.findOne({_id: ObjectId("id")})
 export const getContactById = async (req, res) => {
     try {
-        const contact = await ContactModel.findById(req.params.id);
+        const contact = await ContactModel.findOne({
+            _id: req.params.id,
+            user: req.user.id
+        });
         if (!contact) {
             // 404 HTTP status code for file not found
-            return res.status(404).json({ message: 'Contact not found'});
+            return res.status(404).json({ message: 'Contact not found' });
         }
         res.status(200).json(contact);
     } catch (error) {
@@ -30,7 +33,10 @@ export const getContactById = async (req, res) => {
 // Create a new contact = Same as db.contacts.insertOne()
 export const createContact = async (req, res) => {
     try {
-        const newContact = new ContactModel(req.body);
+        const newContact = new ContactModel({
+            ...req.body,
+            user: req.user.id
+        });
         const savedContact = await newContact.save();
         // 201 HTTP status code for created
         res.status(201).json(savedContact);
@@ -43,14 +49,22 @@ export const createContact = async (req, res) => {
 // Update a contact by ID = Same as db.contacts.updateOne({_id: ObjectId("id")}, {$set: {...}})
 export const updateContact = async (req, res) => {
     try {
-        const updatedContact = await ContactModel.findByIdAndUpdate(req.params.id,req.body, {
-            new: true
+        const contact = await ContactModel.findOne({
+            _id: req.params.id,
+            user: req.user.id
         });
 
-        if (!updatedContact) {
+        if (!contact) {
             // 404 HTTP status code
-            return res.status(404).json({ message: 'Contact not found'});
+            return res.status(404).json({ message: 'Unauthorized access' });
         }
+
+        const updatedContact = await ContactModel.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            { new: true }
+        );
+
         res.status(200).json(updatedContact);
     } catch (error) {
         // 500 HTTP status code for server error
@@ -61,31 +75,51 @@ export const updateContact = async (req, res) => {
 // Delete a contact by ID = same as db.contacts.deleteOne({_id: ObjectId("id")})
 export const deleteContact = async (req, res) => {
     try {
+        const contact = await ContactModel.findOne({
+            _id: req.params.id,
+            user: req.user.id
+        });
+
+        if (!contact) {
+            res.status(403).json({ message: "Unauthorized" })
+            return
+        }
+
         const deletedContact = await ContactModel.findByIdAndDelete(req.params.id);
 
         if (!deletedContact) {
             // 404 HTTP status code
-            return res.status(404).json({ message: 'Contact not found'});
-        } 
+            return res.status(404).json({ message: 'Contact not found' });
+        }
         res.status(200).json({ message: 'Contact deleted successfully' });
-    }  catch (error) {
+    } catch (error) {
         // 500 HTTP status code for server error
         res.status(500).json({ message: error.message });
-    } 
+    }
 }
 
 // Delete all contacts = same as db.contacts.drop()
 export const deleteAllContacts = async (req, res) => {
     try {
+        const contact = await ContactModel.findOne({
+            _id: req.params.id,
+            user: req.user.id
+        });
+
+        if (!contact) {
+            // 404 HTTP status code
+            return res.status(404).json({ message: 'Unauthorized' });
+        }
+
         const deletedContacts = await ContactModel.deleteMany({});
 
         if (deletedContacts.deletedCount === 0) {
             // 404 HTTP status code
-            return res.status(404).json({ message: 'Contacts not found'});
-        } 
+            return res.status(404).json({ message: 'Contacts not found' });
+        }
         res.status(200).json({ message: 'Contacts deleted successfully' });
-    }  catch (error) {
+    } catch (error) {
         // 500 HTTP status code for server error
         res.status(500).json({ message: error.message });
-    } 
+    }
 }
